requirements.txt
ollama
pdfplumber
numpy
paddleocr
pillow
pymupdf
openpyxl